+++ src/listener.js
@@ -76,12 +76,13 @@ OCA.Onlyoffice.onRequestSaveAs = function(saveData) {
 		.catch(() => {})
 }
 
-OCA.Onlyoffice.onRequestInsertImage = function(imageMimes) {
+OCA.Onlyoffice.onRequestInsertImage = function(imageMimes, insertionType) {
 	getFilePickerBuilder(t(OCA.Onlyoffice.AppName, 'Insert image'))
+		.setMultiSelect(true)
 		.setMimeTypeFilter(imageMimes)
 		.addButton({
 			label: t('core', 'Choose'),
-			callback: (nodes) => { if (!nodes[0]) return; document.querySelector(OCA.Onlyoffice.frameSelector).contentWindow.OCA.Onlyoffice.editorInsertImage(nodes[0].path) },
+			callback: (nodes) => { if (!nodes.length) return; document.querySelector(OCA.Onlyoffice.frameSelector).contentWindow.OCA.Onlyoffice.editorInsertImage(nodes.map((node) => node.path), insertionType) },
 			variant: 'primary',
 		})
 		.build()
@@ -192,7 +193,7 @@ window.addEventListener('message', function(event) {
 		OCA.Onlyoffice.onRequestSaveAs(event.data.param)
 		break
 	case 'editorRequestInsertImage':
-		OCA.Onlyoffice.onRequestInsertImage(event.data.param)
+		OCA.Onlyoffice.onRequestInsertImage(event.data.param, event.data.documentSelectionType)
 		break
 	case 'editorRequestSelectSpreadsheet':
 		OCA.Onlyoffice.onRequestSelectSpreadsheet(event.data.param, event.data.documentSelectionType)