+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Changed
+- allow inserting multiple images from Nextcloud into the document at once
 - redesigned the admin settings sections as tabbed navigation
 - changed common and security admin settings to save automatically
 - added supported languages to the Help Center link